refactor: extract monolithic components and store responsibilities to composables
